+++ src/main/java/pl/miloszlewandowski/activities/ActivityRepository.java
@@ -2,6 +2,6 @@ package pl.miloszlewandowski.activities;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivityRepository extends JpaRepository<Activity, Long> {
+public interface ActivityRepository extends JpaRepository<Activity, Integer> {
 
 }
