+++ src/main/java/pl/miloszlewandowski/moodchecks/MoodCheksRepository.java
@@ -2,5 +2,5 @@ package pl.miloszlewandowski.moodchecks;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoodCheksRepository extends JpaRepository<MoodCheck, Long> {
+public interface MoodCheksRepository extends JpaRepository<MoodCheck, Integer> {
 }
