+++ src/main/java/pl/miloszlewandowski/patients/PatientRepository.java
@@ -3,5 +3,5 @@ package pl.miloszlewandowski.patients;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.miloszlewandowski.patients.Patient;
 
-public interface PatientRepository extends JpaRepository<Patient, Long> {
+public interface PatientRepository extends JpaRepository<Patient, Integer> {
 }
