+++ src/main/java/pl/miloszlewandowski/records/RecordCreationDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public class RecordCreationDTO {
 
-    private Long RecordId;
+    private Integer RecordId;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private final LocalDateTime dateTime = LocalDateTime.now();
 }
