+++ src/main/java/pl/miloszlewandowski/records/RecordRepository.java
@@ -3,6 +3,6 @@ package pl.miloszlewandowski.records;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.miloszlewandowski.records.Record;
 
-public interface RecordRepository extends JpaRepository<Record, Long> {
+public interface RecordRepository extends JpaRepository<Record, Integer> {
 //    Page<Record> getAllByPatient(Pageable pageable);
 }
